Huge commit of various work (including unfinished)
Sorry if this breaks things, but is easier to start from a clean
slate. Let me know if your build breaks, and I'll fix it asap.

+++ src/spatial/RangeSet.hh
@@ -12,10 +12,15 @@ namespace spatial
 {
 //===========================================================================
 
+
+/**
+ * TODO: 
+ *   convert from dual  map to rtree
+ */
 class RangeSet
 {
 public:
-  unsigned int size() const
+  size_t size() const
   {
     return ranges.size();
   }
